feat(cloud-picture): fix bsdf glsl(pass compile)

+++ src/run/cloud_picture/domain_layer/domain/shader/ray_tracing/ggx_direct.glsl
@@ -1,4 +1,5 @@
-vec3 computeDirectLight(inout float seed, float epsilon, float tMin,
+
+vec3 computeDirectLight(inout uint seed, float epsilon, float tMin,
                         vec3 worldPosition, vec3 worldNormal, vec3 V,
                         ShadingData shading,
